backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
openai-whisper
pydub
google-cloud-firestore
